refactor: 각 class를 파일별로 분리함
